Fix problem match not finding players in db
The id stored in the db is the same as the slack id of the user.
When a user is mentioned (as in when a match is reported) the string
is of the form <@userId>. We need to get just userId from this to
lookup users in db. I added a utility function for this.